Regularise indentation to use spaces
Currently some tabs seem to be meant to be 2 spaces wide, while other
tabs (sometimes even in the same file, e.g. thexpmap.cxx) seem to be
meant to be 4 spaces wide.

Space-only indentation seems to be dominant in the code, and
standardising on that makes the source easier to work with without
having to set a non-standard tab width in editors.

This patch doesn't eliminate all tab indentation - it's split out of
a "miscellaneous whitespace" patch Wookey has in the Debian packaging
(debian/patches/*therion-5.3.12-whitespace.patch).