Added basic standard implementations for line:fixed-ladder and line:rope-ladder
The code currently draws the side of the ladder in one step (shifted path).
This approach currently has the issue, that when the path moves sideways, the width of the ladder shrinks, because the path shifting is applied directly without respect for angles inside the path.
Multipoint but straight lines are rendered correct most of the time.