Improve calculation of fractional year
Previously this calculated the fraction with:

((month - 1) + (day - 1) / month_length) / 12

Which means the day length varies depending on the month length.  Now
it calculates using a consistent day length (only different between
leap years and non leap years, which is an inherent requirement):

day_of_year / year_length

If we only have a year, the mid point of the year is now used (instead
of the start of the year).

If we only have a year and a month, the 15th of the month is now
used (was the 1st).

If we have the whole date but no time, midday is used (was midnight
at the start of the day).

This is mainly used for calculating the declination, so the intended
effect is that slightly better declination values will be calculated.