Make metapost output reproducible
Set the mpost randomseed so that the same input gives the same output.
The particular motivation is to make the Debian package build
reproducible, and currently it fails to be for the samples.

This seems more widely desirable though - as a user it seems weird
that things like generated rocks change shape each time you generate
a PDF.