Allow building against system libproj
Linux distros dislike use of bundled code copies because it means a lot
more work for security fixes to the bundled code.  The Debian package
of therion currently has a special patch to make it build with the
system libproj.

This change provides a simple way to switch the build use to use the
system libproj instead of the bundled version:

perl ./makeconfig.pl PROJ SYSTEM

And to switch back to using the bundled code copy:

perl ./makeconfig.pl PROJ BUNDLED