xtherion: Remove check for Img package
This commit removes a check for the Img package, which determined the
image file types that would be recognised by XTherion.  This check
appears to be redundant because if the user enables "all file types"
they can select a PNG, furthermore XTherion is able to work with PNGs
despite the Img package not being present.

Closes: #479